Set up initial page structure
The instructions for the project mentioned that this page would basically be seperated out into 4 sections, so I used the standard html boiler plate and for the body set up 4 divs to get the initial structure down. I'm begining the project with the first div, Header, so I began creating the elements that would go in that block. At this time, header logo and the ul are grouped together with the intention of placing space-around the elements, the greeting message is grouped with a header, paragraph, and the sign up button. And lastly the placeholder image to the right. I imagine I could just flex this section and align-self the header-logo div so that it's sitting above and seperate from the other two elements which will be aligned with each other.

The next commit will involve building initial css and moving elements in the header into place. Will also probably integrate the color scheme.