Build Header (In Progress)
I haven't commited the changes in a moment because I've been experimenting with my approach to implementing the layout. Based off the language used on the "colors and font" styles png, I split what was originally the header into two seperate blocks, the header contains the log and links, and the "hero" contains the greeting-text + image. I may group these two boxes together later but, for now, I'm trying to see what I can do without adding 20 flexes to one area of the screen, so I'm trying to use combination selectors to target .header, .hero {}.

I previously forgot to link in the css file so that was fixed here as well.