Complete Top Quadrant
googled how to import the roboto font since it didn't appear to be "standard" or built in. Learned that google fonts has a convient way of linking your html document to where their font is stored as well as the command to call it in css, so I pasted those in.

Still haven't difficulty with arranging things in flexbox, so I looked up another video online (https://www.youtube.com/watch?v=vQAvjof1oe4) and I watched less than 3 minutes of it before I figured out how to solve some problems. Kevin Powell uses > * after a class name as a way to call of the children of that container, then applies flex basis at 100% for equally sized boxes. This allowed me to take the two main boxes in the hero and give them the same area, then I added a gap between them.

The button was quick enough to implement. I remember reading the html on some of the previous lessons and saw that it was easy to add rounded corners, I just had to look it up again to see what the property was called. So far so good, time to start working on the second quandrant.

I do anticipate that there will be compounding issues as I get further down. Ideally I'd prefer that the boxes don't automatically adjust their sizes as I scale the screen and I forgot how to achieve that, so that's on my to do. It just starts to look a bit ugly as you move things around, so I currently set a min-width command to the hero text and the placeholder image square.