Structure upper mid quadrant
Update: #1 I looked at some github projects on lunch and found that some people use the same title over and over again and just specify what itteration they're on. I'll try that for this section.

I've got the HTML portion written and now I'm trying to figure out the flexbox approach for css. My current plan is to set the illustration content to wrap vertically, so only enough space for two blocks will be permitted before it has to wrap around to the right. This should pair each illustration box with its subtext. I also tried to create the boxes themselves with the borders but it's not rendering and I'm not sure why that is yet.